Replaced usage of 'pow' in searches within bitmask-related enums to custom integer-only power() function (in base/mathfunc.h). Rewrote calculation of (sigma / rij)^6 in calculation of LJ vdW terms to avoid pow() function. Changed M_PI to PI in render/objects.cpp. Reordered namespace in parse/parser.h. Added 'define NOMINMAX' to templates to prevent macro errors when compiling on Windows machines. Removed reference to round() in Prefs::estimateEwaldParameters() with floor(x+0.5). Removed problematic (for Windows) automatic variables / arrays in Model::selectionEmpirical(), FourierData::prepare(), Ewald reciprocal energy / force calculation, and Canvas::endMode(). Removed version from specfile name. Modified changeversion script to leave svn well alone!

+++ src/parse/parser.cpp
@@ -30,7 +30,8 @@ Parser parser;
 const char *ParseOptionKeywords[Parser::nParseOptions] = { "defaults", "usequotes", "skipblanks", "stripbrackets", "__DUMMY__" };
 Parser::ParseOption Parser::parseOption(const char *s)
 {
-	return (Parser::ParseOption) int(pow(2,enumSearch("parse option", Parser::nParseOptions, ParseOptionKeywords, s))); }
+	return (Parser::ParseOption) pow(2,enumSearch("parse option", Parser::nParseOptions, ParseOptionKeywords, s));
+}
 
 // Constructor
 Parser::Parser()
